// src/app/pages/models/braking-system.model.ts
export class BrakingSystemOfVehicle {
  brakingSystemId: number;
  name: string;
}
export class BrakingSystem {
  brakingSystemId: number;
  name: string;
  description: string;
  active: string;
  createdAt: string;
  updatedAt: string;
}
export class BrakingSystemDetails {
  brakingSystemId: number;
  name: string;
  description: string;
  active: string;
  createdAt: string;
  updatedAt: string;
  createdBy: number;
  updatedBy: number;
}