// src/app/pages/models/insurance-company.model.ts
export class InsuranceCompany {
    insuranceCompanyId: number;
    name: string;
    description: string;
    active: string;
    createdAt: string;
    updatedAt: string;
}
export class InsuranceCompanyDetails {
    insuranceCompanyId: number;
    name: string;
    description: string;
    active: string;
    createdAt: string;
    updatedAt: string;
    createdBy: number;
    updatedBy: number;
}