// src/app/pages/models/insurance-type.model.ts
export class InsuranceType {
    insuranceTypeId: number;
    type: string;
    active: string;
    createdAt: string;
    updatedAt: string;
}
export class InsuranceTypeDetails {
    insuranceTypeId: number;
    type: string;
    active: string;
    createdAt: string;
    updatedAt: string;
    createdBy: number;
    updatedBy: number;
}