// src/app/pages/models/vehicle-color.model.ts
export class ColorOfVehicle {
    vehicleColorId: number;
    color: string;
}
export class VehicleColor {
    vehicleColorId: number;
    color: string;
    active: string;
    createdAt: string;
    updatedAt: string;
}
export class VehicleColorDetails {
    vehicleColorId: number;
    color: string;
    active: string;
    createdAt: string;
    updatedAt: string;
    createdBy: number;
    updatedBy: number;
}