// src/app/pages/models/vehicle-condition.model.ts
export class VehicleCondition {
    vehicleConditionId: number;
    condition: string;
    active: string;
    createdAt: string;
    updatedAt: string;
}
export class VehicleConditionDetails {
    vehicleConditionId: number;
    condition: string;
    active: string;
    createdAt: string;
    updatedAt: string;
    createdBy: number;
    updatedBy: number;
}