// src/app/pages/models/vehicle-variant.model.ts
export class VariantOfVehicle {
    vehicleVariantId: number;
    variant: string;
}
export class VehicleVariant {
    vehicleVariantId: number;
    variant: string;
    active: string;
    createdAt: string;
    updatedAt: string;
}
export class VehicleVariantDetails {
    vehicleVariantId: number;
    variant: string;
    active: string;
    createdAt: string;
    updatedAt: string;
    createdBy: number;
    updatedBy: number;
}